add entity service and controller to restaurant service

+++ restaurant-service/src/main/java/com/merveyilmaz/restaurantservice/dto/RestaurantDTO.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 public record RestaurantDTO(String id,
                             String name,
                             String restaurantCreateDate,
-                            double longitude,
-                            double latitude) {
+                            int longitude,
+                            int latitude) {
 }
