+++ restaurant-service/src/main/java/com/merveyilmaz/restaurantservice/request/RestaurantSaveRequest.java
@@ -4,6 +4,6 @@ package com.merveyilmaz.restaurantservice.request;
 public record RestaurantSaveRequest(String id,
                                     String name,
                                     String restaurantCreateDate,
-                                    double longitude,
-                                    double latitude) {
+                                    int longitude,
+                                    int latitude) {
 }
